SoundGate/requirements.txt
pyserial
pandas
dash
dash-bootstrap-components